Evaluate and document variance implementation differences between area() and volume()
- Fixed decimal type error in area() variance calculation (cast EXPNS to Float64)
- Fixed missing columns issue in volume() get_cond_columns() method
- Created comprehensive variance implementation analysis and proposal
- Added test scripts comparing variance behavior between functions
- Documented that area() implements proper stratified variance while volume() uses 12% placeholder

Key findings:
- area(): Proper domain total estimation variance following Bechtold & Patterson (2005)
- volume(): Simplified 12% CV placeholder, needs ratio-of-means variance implementation
- Proposal includes detailed implementation plan for improving volume() variance

Tests show area() CV responds to sample size and domain restrictions (0.85% for Texas),
while volume() remains fixed at 12% regardless of conditions.

+++ src/pyfia/estimation/estimators/area.py
@@ -510,7 +510,8 @@ class AreaEstimator(BaseEstimator):
 
         variance_components = strata_stats.with_columns([
             # Multiply by n_h, not divide by it!
-            (pl.col("w_h") ** 2 * pl.col("s2_yh") * pl.col("n_h")).alias("v_h")
+            # Cast w_h to Float64 to handle decimal types from database
+            (pl.col("w_h").cast(pl.Float64) ** 2 * pl.col("s2_yh") * pl.col("n_h")).alias("v_h")
         ])
 
         # Step 5: Sum variance components
